feat: refactor parameter
