+++ src/PayloadTemplates/WaitForTokenPayloadTemplate.ts
@@ -1,1 +1,11 @@
-export abstract class WaitFotTokenPayloadTemplate {}
+export abstract class WaitFotTokenPayloadTemplate {
+  abstract process(inputJson: string): Record<string, unknown>;
+
+  protected isPathKey(path: string): boolean {
+    return path.endsWith('.$');
+  }
+
+  protected isContextObjectPath(path: string): boolean {
+    return path.startsWith('$$.');
+  }
+}
